[config] storage_config own file
per title

Closes: #1722
Approved by: mimoo
